better work with AtomicLong (replace it with Int in some places) + change vars to vals

+++ src/main/scala/org/myapp/EventStorage.scala
@@ -9,7 +9,7 @@ trait EventStorage[F[_]] {
 
   def addEvent(event: Event): F[Unit]
 
-  def updateEvent(id: AtomicLong, event: Event): F[Unit]
+  def updateEvent(id: Int, event: Event): F[Unit]
 
 
 }
